sources.cfg: always clone with https.
The problem is that the kgs' sources.cfg includes private repositories, now using the cloneurl option.
Private repositories require authentication (not ssh-key) when cloned using https.
plonesource.org always uses https for cloning, but that only works for public repositories.
Therefore we change the cloneurl to ssh for all repositories.

/cc @maethu