Introduce github-private-cloneurl.
This reverts 2940de12, which undermines the cloneurl functionality of plonesource.org.
Public repositories should **always** be cloned with https, not ssh.
There are servers which disallow ssh connections to github.

This change introduces a ``github-private-cloneurl``, which defaults to ``ssh`` and
is used for private repositories (added by the kgs sources) without changing the
cloneurl for public repositories.